Start Fix #25, BE Statistik: Rücksetzung nur für bestimmte Gruppen

+++ src/Resources/contao/modules/ModuleDlstatsStatistics.php
@@ -46,6 +46,13 @@ class ModuleDlstatsStatistics extends \Contao\BackendModule
     protected $username   = '---';
 
     /**
+	 * is user allowed to reset the statistic
+	 * @var bool
+	 */
+    protected $boolAllowReset = true;
+    
+
+    /**
      * Constructor
      */
     public function __construct()
@@ -61,6 +68,13 @@ class ModuleDlstatsStatistics extends \Contao\BackendModule
          && (int) ($GLOBALS['TL_CONFIG']['dlstatLastDownloads']) >0)
         {
             $this->intLastDownloadLimit = (int) ($GLOBALS['TL_CONFIG']['dlstatLastDownloads']);
+        }
+
+        if (isset($GLOBALS['TL_CONFIG']['dlstatStatresetProtected'])
+         && isset($GLOBALS['TL_CONFIG']['dlstatStatresetGroups'])
+         && (int) ($GLOBALS['TL_CONFIG']['dlstatStatresetProtected']) >0)
+        {
+            $this->boolAllowReset = $this->isUserInDownloadStatGroups($GLOBALS['TL_CONFIG']['dlstatStatresetGroups']);
         }
 
         if (\Input::get('act', true)=='zero')
@@ -129,6 +143,8 @@ class ModuleDlstatsStatistics extends \Contao\BackendModule
             $this->Template->alldownloads = $this->getAllDownloadsFiltered();
         }
 
+        $this->Template->allow_reset      = $this->boolAllowReset;
+
     }
 
     /**
@@ -136,6 +152,10 @@ class ModuleDlstatsStatistics extends \Contao\BackendModule
      */
     protected function setZero()
     {
+        if (false === $this->boolAllowReset)
+	    {
+	        return ;
+	    }
         \Database::getInstance()->prepare("TRUNCATE TABLE tl_dlstatdets")->execute();
         \Database::getInstance()->prepare("TRUNCATE TABLE tl_dlstats")->execute();
 
@@ -605,4 +625,38 @@ class ModuleDlstatsStatistics extends \Contao\BackendModule
 
         return $AllDownloads;
     }
+
+    /**
+	 * Check if User member of group in downlaod statistik groups for reset
+	 *
+	 * @param   string  serialized array
+	 * @return  bool    true / false
+	 */
+	protected function isUserInDownloadStatGroups($dlstats_reset_groups)
+	{
+	    if ( true === $this->User->isAdmin )
+	    {
+	        //Debug log_message('Ich bin Admin', 'dlstats_debug.log');
+	        return true; // Admin darf immer
+	    }
+	    
+	    //Schutz aktiviert, Einschränkungen vorhanden?
+	    if (0 == strlen($dlstats_reset_groups))
+	    {
+	        //Debug log_message('dlstats_stat_groups ist leer', 'dlstats_debug.log');
+	        return false; // nicht gefiltert, also darf keiner außer Admin
+	    }
+	     
+	    //mit isMemberOf ermitteln, ob user Member einer der erlaubten Gruppen ist
+	    foreach (\Contao\StringUtil::deserialize($dlstats_reset_groups) as $id => $groupid)
+	    {
+	        if ( true === $this->User->isMemberOf($groupid) )
+	        {
+	            //Debug log_message('Ich bin in der richtigen Gruppe', 'dlstats_debug.log');
+	            return true; // User is Member of allowed groups
+	        }
+	    }
+	    //Debug log_message('Ich bin in der falschen Gruppe', 'dlstats_debug.log');
+	    return false;
+	}
 }
